--- jupyterlab_react/src/style.ts
@@ -1,28 +1,0 @@
-const style = {
-		container: {
-			padding: "20px"
-		},
-		button: {
-			background: "#eee",
-			color: "#444",
-			borderRadius: "5px",
-			marginTop: "5px",
-			marginBottom: "10px",
-			marginRight: "4px",
-			marginLeft: "4px"
-		},
-		wrapper: {
-			paddingTop: "20px",
-			paddingLeft: "20px",
-		},
-		hr: {
-			height: "1px",
-			marginTop: "10px",
-			marginBottom: "10px"
-		},
-		leftAligned: {
-			textAlign: "left" as const
-		}
-	}
-
-export {style}
